backend/requirements.txt
asgiref==3.2.3
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
Django==3.0.2
django-environ==0.4.5
django-extensions==2.2.6
django-redis==4.11.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
inflection==0.5.0
ipython==7.11.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.15.2
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
packaging==20.4
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.7
pytz==2019.3
raven==6.10.0
redis==3.3.11
requests==2.22.0
rope==0.16.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplegeneric==0.8.1
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.0
